+++ src/Server.cpp
@@ -121,63 +121,6 @@ void Server::chat_in(int fd)
 		build_in(fd, cmds[i]);
 }
 
-static bool is_enough_params(Client &client, const Command& cmd)
-{
-	if (cmd._params.size() != 2)
-	{
-		send_errmsg(client, 461, cmd.get_original_str() + " :Not enough parameters");
-		return false;
-	}
-	return true;
-}
-
-static bool is_exist_channel(Client&client, const std::string &channel, ChannelManager& chm)
-{
-	if (!chm.exist_channel(channel))
-	{
-		send_errmsg(client, 403, channel + " :No such channel");
-		return false;
-	}	
-	return true;
-}
-
-static bool is_exist_user(Client&client, const std::string &channel, ChannelManager& chm)
-{
-	if (!chm.exist_channel(channel))
-	{
-		send_errmsg(client, 403, channel + " :No such channel");
-		return false;
-	}	
-	return true;
-}
-
-static bool is_exist_user(Client &client, const std::string& kick_user, ClientManager& clientManager)
-{
-	if (!clientManager.exist_client_by_nick(kick_user))
-	{
-		send_errmsg(client, 401, kick_user + " :No such nick");
-		return false;
-	}
-	return true;
-}
-#include "ChannelManager.hpp"
-#include "CheckRegister.hpp"
-
-void kick(Client &client, const Command& cmd, ClientManager& client_manager, ChannelManager& channel_manager)
-{
-	if (!is_authenticated(client)) return;
-	if (!is_seted_nick_user(client)) return;
-	if (!is_enough_params(client, cmd))	return;
-	std::string channel_name = cmd._params[0];
-	std::string kick_user_name = cmd._params[1];
-	if (!is_exist_channel(client, channel_name, channel_manager))  return;
-	if (!is_exist_user(client, kick_user_name, client_manager))  return;
-	Channel& channel = channel_manager.get_channel(channel_name);
-	Client &kick_user = client_manager.get_client_by_nick(kick_user_name);
-	std::string kick_reason = cmd._params[0];
-	channel.kick(client, kick_user, kick_reason);
-}
-
 void Server::build_in(int fd, const Command &cmd)
 {
 	Client &client = clientManager.get_client_by_fd(fd);
@@ -221,7 +164,6 @@ void Server::build_in(int fd, const Command &cmd)
 			break;
 		case KICK:
 			kick(client, cmd, clientManager, channelManager);
-			std::cout << "kick" << std::endl;
 			break;
 		case INVITE:
 			std::cout << "invite" << std::endl;
